Move nested types to separate files
