+++ src/Features/LanguageServer/Protocol/Handler/Diagnostics/DiagnosticsRefreshQueue.Factory.cs
@@ -1,0 +1,41 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using System.Composition;
+using Microsoft.CodeAnalysis.Host.Mef;
+using Microsoft.CodeAnalysis.Shared.TestHooks;
+
+namespace Microsoft.CodeAnalysis.LanguageServer.Handler.Diagnostics;
+
+internal sealed partial class DiagnosticsRefreshQueue : AbstractRefreshQueue
+{
+    [ExportCSharpVisualBasicLspServiceFactory(typeof(DiagnosticsRefreshQueue)), Shared]
+    internal sealed class Factory : ILspServiceFactory
+    {
+        private readonly IAsynchronousOperationListenerProvider _asyncListenerProvider;
+        private readonly LspWorkspaceRegistrationService _lspWorkspaceRegistrationService;
+        private readonly Refresher _refresher;
+
+        [ImportingConstructor]
+        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
+        public Factory(
+            IAsynchronousOperationListenerProvider asynchronousOperationListenerProvider,
+            LspWorkspaceRegistrationService lspWorkspaceRegistrationService,
+            Refresher refresher)
+        {
+            _asyncListenerProvider = asynchronousOperationListenerProvider;
+            _lspWorkspaceRegistrationService = lspWorkspaceRegistrationService;
+            _refresher = refresher;
+        }
+
+        public ILspService CreateILspService(LspServices lspServices, WellKnownLspServerKinds serverKind)
+        {
+            var notificationManager = lspServices.GetRequiredService<IClientLanguageServerManager>();
+            var lspWorkspaceManager = lspServices.GetRequiredService<LspWorkspaceManager>();
+
+            return new DiagnosticsRefreshQueue(_asyncListenerProvider, _lspWorkspaceRegistrationService, lspWorkspaceManager, notificationManager, _refresher);
+        }
+    }
+}
